requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
openai
python-dotenv
